feat: added derivation path to account data
