+++ src/broker/Background.js
@@ -1,4 +1,4 @@
-import { ChainNetworks } from '../store/utils'
+import { ChainNetworks } from '@/utils/networks'
 import buildConfig from '../build.config'
 import { BG_PREFIX, handleConnection, removeConnectId, getRootURL } from './utils'
 import { assets } from '@liquality/cryptoassets'
