+++ src/contentScript.js
@@ -3,7 +3,7 @@ import { inject } from './broker/utils'
 import Script from './broker/Script'
 import { providerManager, ethereumProvider, overrideEthereum, bitcoinProvider, nearProvider, paymentUriHandler, solanaProvider } from './inject'
 import buildConfig from './build.config'
-import { ChainNetworks } from './store/utils'
+import { ChainNetworks } from '@/utils/networks'
 import { chains, isEthereumChain } from '@liquality/cryptoassets'
 
 ;(new Script()).start()
