+++ src/swaps/oneinch/OneinchSwapProvider.js
@@ -3,7 +3,7 @@ import BN from 'bignumber.js'
 import { SwapProvider } from '../SwapProvider'
 import { v4 as uuidv4 } from 'uuid'
 import { chains, currencyToUnit, unitToCurrency } from '@liquality/cryptoassets'
-import { ChainNetworks } from '../../store/utils'
+import { ChainNetworks } from '@/utils/networks'
 import { withLock, withInterval } from '../../store/actions/performNextAction/utils'
 import { prettyBalance } from '../../utils/coinFormatter'
 import { isEthereumChain, isERC20 } from '@/utils/asset'
