+++ src/utils/accounts.js
@@ -9,6 +9,7 @@ export const accountCreator = (payload) => {
     chain,
     index,
     addresses,
+    derivationPath,
     assets,
     balances,
     type,
@@ -29,6 +30,7 @@ export const accountCreator = (payload) => {
     name,
     chain,
     index,
+    derivationPath,
     addresses: _addresses,
     assets,
     balances: balances || {},
